+++ CHANGELOG.md
@@ -398,6 +398,7 @@
 - Added `q.has_argument(<argument_name>, [argument_value])` MoLang function to several battle-related action effect contexts.
 - Added `q.has_argument_at(<index>, [argument_value])` MoLang function to several battle-related action effect contexts.
 - Added `q.hit_count` MoLang function to move action effect contexts.
+- Added `is_included`, `to_lower`, `to_upper`, and `string_length` as available Molang functions.
 
 ## [1.6.1 (January 26th, 2025)](#1-6-1)